Start tackling some bugs in ExpandingListView
I know I don't HAVE to do this, but we're already using this code, and it works
pretty darn well. It's just that the Googler who wrote it left some pretty sad
little bugs in it, and I've spent a few hours tracking down just what the
causes of these bugs are, and at least formulating action plans to fix them.