General cleanup, import organizing. Create KiwiDoneCancelActivity for easy design.
Might I be over-engineering a solution? Yes. Will my over-engineering help us
overall? Yes.

The idea behind the KiwiDoneCancelActivity is to abstract away all the
necessary code that goes into making the activity have the Done/Cancel buttons
(or just the Done button with Cancel in the menu) and lets us just focus on
building the interactions with the UI.

I've made it so that it should be very easy to make a new activity that has
Done|Cancel in the action bar (literally just need to override two methods:
getLayoutResource and getDoneCancelHandler) or to make an activity with the
Done button and Cancel in the menu (override four methods: getLayoutResource,
hasCancelButton, getDoneButtonHandler and getCancelFromMenuHandler). Anything
that has to happen in the onCreate method can be done in the onCreated method
(I made the onCreate method final in this new abstract class because there is
some special logic that goes on in there and I don't want us to accidentally
forget to call super.onCreate(savedInstanceState))

While I was at it, I also built up some infrastructure to allow fragments, or
anything really, to hook into the Done/Cancel interaction of a
KiwiDoneCancelActivity, so they can be notified if the user clicks Done or
Cancel (most important is Done), and they can even say "Hey, you can't be done
yet!" if need be. An example of a scenario where this will be useful is on an
Add Assignment page. The EditAssignmentFragment is entirely responsible for
handling user input, etc. in the form, and it may as well handle the database
interactions while we're at it. We would either have to have the
AddAssignmentActivity be intimately aware that the EditAssignmentFragment lives
inside it, and then call a method on that fragment to say "I'm clicking Done,
is that okay?", ... or we could build it the other way around, where the
EditAssignmentFragment is aware that it's in an activity that has Done/Cancel
functionality, and is able to register for that information. I personally feel
that this design is better, especially when you consider what we'll be doing on
the page after selecting the courses to import -- presenting a ViewPager with
multiple EditCourseFragments, but with only one Done button. If the
EditCourseFragments each sign up to know when the user clicks Done, and then
are able to say "no, wait, you need to fix this", it's way easier to implement.

--- The only issue there is that on that multiple course edit page, how do we
go back to those fragments, after we know they're all valid, and say "okay, you
can save now"? That will take some thinking.