Override getCount in DummyItemAdapter.
Without this, getCount would always return one less than we actually want
(because the super class doesn't know about the dummy item). We can't safely
shove the dummy item at the front of the objects list (otherwise we'd be
modifying the list object passed in, unbeknownst to users of this code). So, we
just override getCount() to return the size of the 'objects' list, plus 1 for
the dummy item.