Add new DbAndCursor class to DatabaseHandler.
Android will complain if we naively keep the database connections open (because
we leak those connections if the screen is rotated, etc.). For DatabaseHandler
methods that previously returned a Cursor (i.e. getRawAssignmentCursor) we need
to instead return both the Cursor and the SQLiteDatabase connection it's using.
(Alright, we could actually hide the database inside this DbAndCursor object,
and just expose a close() method for it... Might want to keep that in mind for
the future.) In OverviewFragment's onDestroy method, we need to close this
database object.