So many changes I don't even know what to say.
I added a custom view to display in the overview fragment if the assignment
list is empty, and even added logic to decide what to show inside there
depending on if there are really no assignments, or just nothing to show.
(Specifically, this decides whether the filter button [in both the custom view
and the menu] should be visible.)

I actually got the EditAssignmentActivity integrated into being able to save
assignments. Haven't quite gotten it to fill in the fields with the assignment
data, but that ain't hard.

Cleaned up some imports, other code, added logic to some miscellaneous
functions to avoid or at least log obscene memory usage.