Make sections in FilterDialogFragment expandable/collapsible.
Some parts of this are a bit janky (especially expanding the types section) but
it does what we wanted.

One thing we could fix, but won't focus on now, is that if you rotate the
screen after expanding some sections, the layout just restarts with all
sections closed. We could save the state and restore it, but eh.

Closes #21